# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rahul Raj — Portfolio & Systems Architecture API",
    "version": "1.0.0",
    "description": "Production-grade RESTful API documentation for Rahul Raj's modern portfolio platform, built with **Next.js 16**, **TypeScript**, **MongoDB Atlas**, **Cloudinary**, and **AI LLM Orchestration**.\n\n### Key Architectural Pillars\n- **Public Gateway**: Ultra-fast endpoints for portfolio contents, live ATS resume downloads, real-time presence heartbeats, and AI assistant interaction.\n- **Admin Control Studio & Hardware Biometrics**: Role-Based Access Control (RBAC) secured via WebAuthn/FIDO2 hardware biometric passkeys (Fingerprint, TouchID, Windows Hello, FaceID), 4-digit security PIN, master passcode, cryptographic HMAC tokens, or Email OTP resets.\n- **Resilience Engine**: Zero-crash persistence architecture backed by MongoDB Atlas with in-memory caching and real-time health telemetry probes.\n- **Security & Anti-Abuse**: IP-based sliding window rate limiters, anti-bot honeypot traps, and encrypted session cookies.",
    "contact": {
      "name": "Rahul Raj Modi",
      "email": "rahulraj21480@gmail.com",
      "url": "https://github.com/Rahul-2148"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Environment / Origin Server"
    },
    {
      "url": "http://localhost:3000",
...
    }
  ],
...
  "paths": {
...
    "/api/admin/achievements": {
      "get": {
        "tags": [
          "Admin CMS Content"
        ],
        "summary": "List achievements and certifications",
        "security": [
          {
            "AdminCookieAuth": []
          },
          {
            "AdminPasscodeAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of achievements"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Admin CMS Content"
        ],
        "summary": "Create or modify achievements",
        "security": [
          {
            "AdminCookieAuth": []
          },
          {
            "AdminPasscodeAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Achievement saved"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AdminCookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "portfolio_admin_token",
        "description": "Encrypted HTTP-only admin session token generated during login"
      },
      "AdminPasscodeAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-passcode",
        "description": "Secret master passcode configured via ADMIN_PASSCODE environment variable"
      }
    },
...
  }
}